+++ dashboard/apps/dashboard/admin.py
@@ -1,29 +1,17 @@
 from datetime import date
-import re
-
 from django.conf import settings
 from django.conf.urls import url
 from django.contrib import admin
 from django.contrib.admin.models import LogEntry
-from django.contrib.auth.admin import csrf_protect_m
-from django.contrib.auth.decorators import permission_required
-from django.core.checks import messages
 from django.core import urlresolvers
 from django.core.exceptions import PermissionDenied
-from django.db import transaction
 from django.http import HttpResponse
-from django.shortcuts import render_to_response
-from django.template import RequestContext
-from django.utils.decorators import method_decorator
 from django.utils.html import format_html
 
-from dateutil.relativedelta import relativedelta
-from dashboard.libs.rate_converter import last_date_in_month
-
+from dashboard.apps.dashboard.spreadsheets import Export
 from .models import (Person, Rate, Area, Product, Cost, Budget,
                      ProductStatus, ProductGroupStatus, Saving, Link,
                      PersonCost, Department, Skill)
-from .forms import (PayrollUploadForm, ExportForm, Export)
 from .permissions import ReadOnlyPermissions, FinancePermissions
 from .filters import (
     IsVisibleFilter, IsCivilServantFilter, IsCurrentStaffFilter, HasRateFilter)
@@ -84,74 +72,19 @@ class PersonAdmin(ReadOnlyAdmin, FinancePermissions):
     get_skills.allow_tags = True
 
     def get_urls(self):
-        urls = [
-            url(
-                r'^upload/$',
-                self.admin_site.admin_view(self.upoload_view),
-                name='person_upload_payroll'),
+        urls = super(PersonAdmin, self).get_urls()
+        urls += [
             url(
                 r'^export_rates/$',
                 self.admin_site.admin_view(self.export_person_rates_view),
                 name='person_export_rates'),
         ]
-        return urls + super(PersonAdmin, self).get_urls()
+        return urls
 
-    def has_upload_permission(self, request):
-        return self.is_finance(request.user)
-
-    def get_model_perms(self, request):
-        perms = super(PersonAdmin, self).get_model_perms(request)
-        perms.update({
-            'upload': self.has_upload_permission(request),
-        })
-        return perms
-
-    @csrf_protect_m
-    @transaction.atomic
-    @method_decorator(permission_required('dashboard.upload_person',
-                                          raise_exception=True))
-    def upoload_view(self, request, *args, **kwargs):
-        if not self.has_upload_permission(request):
-            raise PermissionDenied
-
-        start = (date.today() - relativedelta(months=1)).replace(day=1)
-        end = last_date_in_month(start)
-
-        initial = {
-            'date_range': (start, end)
-        }
-
-        if request.method == 'POST':
-            form = PayrollUploadForm(data=request.POST, files=request.FILES,
-                                     initial=initial)
-            if form.is_valid():
-                form.save()
-                level = messages.INFO
-                message = 'Successfully uploaded %s payroll' % form.month
-            else:
-                level = messages.ERROR
-                message = 'Errors uploading %s payroll' % form.month
-
-            self.message_user(request, message, level=level)
-        else:
-            form = PayrollUploadForm(initial=initial)
-
-        context = self.admin_site.each_context(request)
-        context.update({
-            'opts': self.model._meta,
-            'has_permission': self.is_finance(request.user),
-            'form': form,
-        })
-
-        return render_to_response(
-            'admin/dashboard/upload.html',
-            context,
-            context_instance=RequestContext(request))
-
-    @method_decorator(permission_required('dashboard.export_person_rates',
-                                          raise_exception=True))
+    # @method_decorator(permission_required('dashboard.export_person_rates',
+    #                                       raise_exception=True))
     def export_person_rates_view(self, request, *args, **kwargs):
-        if not self.has_upload_permission(request):
+        if not self.is_finance(request.user):
             raise PermissionDenied
         today = date.today()
         export = Export({'date': today, 'title': 'Rate Export'})
@@ -264,66 +197,6 @@ class ProductAdmin(admin.ModelAdmin, FinancePermissions):
                            float_id=obj.float_id)
     float_link.short_description = 'Float Id'
     float_link.allow_tags = True
-
-    def get_urls(self):
-        urls = [
-            url(
-                r'^export/$',
-                self.admin_site.admin_view(self.export_view),
-                name='product_export'),
-        ]
-        return urls + super(ProductAdmin, self).get_urls()
-
-    @csrf_protect_m
-    @transaction.atomic
-    @method_decorator(permission_required('dashboard.adjustmentexport_product',
-                                          raise_exception=True))
-    def export_view(self, request):
-        if not self.is_finance(request.user):  # pragma: no cover
-            raise PermissionDenied
-
-        start = (date.today() - relativedelta(months=1)).replace(day=1)
-        end = last_date_in_month(start)
-
-        initial = {
-            'date_range': (start, end)
-        }
-
-        if request.method == 'POST':
-            form = ExportForm(data=request.POST, files=request.FILES,
-                              initial=initial)
-            if form.is_valid():
-                fname = '%s_%s_%s-%s.xlsm' % (
-                    form.cleaned_data['export_type'],
-                    re.sub(
-                        '[^0-9a-zA-Z]+',
-                        '-',
-                        form.cleaned_data['product'].name),
-                    form.cleaned_data['date_range'][0].year,
-                    form.cleaned_data['date_range'][0].month)
-                workbook = form.export()
-                response = HttpResponse(
-                    content_type="application/vnd.ms-excel")
-                response['Content-Disposition'] = 'attachment; filename=%s' \
-                                                  % fname
-                workbook.save(response)
-                return response
-        else:
-            form = ExportForm(initial=initial)
-
-        context = self.admin_site.each_context(request)
-        context.update({
-            'opts': self.model._meta,
-            'has_permission': self.is_finance(request.user),
-            'form': form,
-            'areas': Area.objects.exclude(products__isnull=True)
-                        .order_by('name').prefetch_related('products'),
-        })
-
-        return render_to_response(
-            'admin/dashboard/export.html',
-            context,
-            context_instance=RequestContext(request))
 
 
 class TaskAdmin(ReadOnlyAdmin):
