+++ dashboard/apps/dashboard/fixtures/auth_group_permissions.yaml
@@ -23,13 +23,13 @@
       - ['change_area', 'dashboard', 'area']
       - ['change_department', 'dashboard', 'department']
       - ['change_person', 'dashboard', 'person']
-      - ['upload_person', 'dashboard', 'person']
       - ['change_skill', 'dashboard', 'skill']
-      - ['export_person_rates', 'dashboard', 'person']
       - ['change_product', 'dashboard', 'product']
-      - ['adjustmentexport_product', 'dashboard', 'product']
-      - ['intercompanyexport_product', 'dashboard', 'product']
-      - ['productdetailexport_product', 'dashboard', 'product']
+      - ['export_person_rates', 'dashboard', 'person']
+      - ['upload_payroll', 'reports', 'productproxy']
+      - ['adjustment_export', 'reports', 'productproxy']
+      - ['intercompany_export', 'reports', 'productproxy']
+      - ['productdetail_export', 'reports', 'productproxy']
       - ['change_task', 'dashboard', 'task']
       - ['change_cost', 'dashboard', 'cost']
       - ['add_cost', 'dashboard', 'cost']
