+++ dashboard/templates/admin/partials/_dashboard_admin_nav.html
@@ -11,9 +11,9 @@
     {% endif %}
     {% if request.user.is_finance %}
       {% url 'admin:person_upload_payroll' as upload_url %}
-      <li><a href="{{ upload_url }}" class="{% if request.path == upload_url %} active{% endif %}">Salary upload</a></li>
+      {%  if upload_url %}<li><a href="{{ upload_url }}" class="{% if request.path == upload_url %} active{% endif %}">Salary upload</a></li>{% endif %}
       {% url 'admin:product_export' as export_url %}
-      <li><a href="{{ export_url }}" class="{% if request.path == export_url %} active{% endif %}">Export</a></li>
+      {%  if export_url %}<li><a href="{{ export_url }}" class="{% if request.path == export_url %} active{% endif %}">Export</a></li>{% endif %}
     {% endif %}
   </ul>
 </div>
